requirements.txt
fastapi~=0.111.1
uvicorn
jinja2
openai~=1.35.15
python-multipart
email-validator

pydantic~=2.8.2
python-dotenv~=1.0.1
twilio~=9.2.3